+++ api/src/main/java/br/com/academiadev/bluerefund/service/AdminService.java
@@ -1,8 +1,5 @@
 package br.com.academiadev.bluerefund.service;
 
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
 import org.springframework.beans.factory.annotation.Autowired;
 
 import br.com.academiadev.bluerefund.model.Admin;
@@ -27,18 +24,5 @@ public class AdminService {
 		adminRepository.save(admin);
 	}
 	
-	public boolean validarEmail(String email)
-    {
-        boolean isEmailIdValid = false;
-        if (email != null && email.length() > 0) {
-            String expression = "^[\\w\\.-]+@([\\w\\-]+\\.)+[A-Z]{2,4}$";
-            Pattern pattern = Pattern.compile(expression, Pattern.CASE_INSENSITIVE);
-            Matcher matcher = pattern.matcher(email);
-            if (matcher.matches()) {
-                isEmailIdValid = true;
-            }
-        }
-        return isEmailIdValid;
-    }
 
 }
