+++ api/src/main/java/br/com/academiadev/bluerefund/service/EmailService.java
@@ -1,0 +1,22 @@
+package br.com.academiadev.bluerefund.service;
+
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+public class EmailService {
+	
+	public boolean validarEmail(String email)
+    {
+        boolean isEmailIdValid = false;
+        if (email != null && email.length() > 0) {
+            String expression = "^[\\w\\.-]+@([\\w\\-]+\\.)+[A-Z]{2,4}$";
+            Pattern pattern = Pattern.compile(expression, Pattern.CASE_INSENSITIVE);
+            Matcher matcher = pattern.matcher(email);
+            if (matcher.matches()) {
+                isEmailIdValid = true;
+            }
+        }
+        return isEmailIdValid;
+    }
+
+}
