+++ api/src/main/java/br/com/academiadev/bluerefund/service/EmpresaInterface.java
@@ -1,0 +1,7 @@
+package br.com.academiadev.bluerefund.service;
+
+public interface EmpresaInterface {
+
+	public void cadastrar(String nome);
+	
+}
